requirements.txt
gdown==5.2.0
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
openai==1.52.0
pandas==2.2.2
streamlit==1.38.0
